UC confessions
This is a small Android app that I wrote while I was learning to program.
It runs on an older version of Android.

Each University of California is allowed to post. Any other UC is allowed
to view a post from another UC. Posts are simple text messeges updating
other schools.